Renamed directory locale to localization
