+++ setup.py
@@ -39,12 +39,12 @@ PACKAGE_DATA = {
 }
 
 my_list = []
-for curr_dir, _, files in os.walk('./src/robotide/locale'):
+for curr_dir, _, files in os.walk('src/robotide/localization'):
     for item in files:
         if '.' in item:
-             my_list.append(os.path.join(curr_dir, item).replace('\\', '/').replace('./src/robotide/locale/', ''))
+             my_list.append(os.path.join(curr_dir, item).replace('\\', '/').replace('./src/robotide/localization/', ''))
 
-PACKAGE_DATA['robotide.locale'] = my_list
+PACKAGE_DATA['robotide.localization'] = my_list
 
 print(f"DEBUG: setup.py  {PACKAGE_DATA=}")
