+++ src/robotide/localization/RIDE.pot
