+++ src/robotide/localization/TRANSLATORS.adoc
