+++ src/robotide/localization/pt_BR/LC_MESSAGES/RIDE.mo
