+++ src/robotide/localization/pt_PT/LC_MESSAGES/RIDE.mo
